Refactoring.
