+++ Src/Application/DependencyInjection.cs
@@ -2,7 +2,7 @@
 using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
-using Northwind.Application.Infrastructure;
+using Northwind.Application.Infrastructure.Behaviours;
 
 namespace Northwind.Application
 {
